Refactor views into separate modules
